+++ circonus/resource_circonus_metric_test.go
@@ -24,26 +24,15 @@ func TestAccCirconusMetric_basic(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_average", "name", metricAvgName),
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_average", "active", "false"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_average", "tags.#", "2"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_average", "tags.2087084518", "author:terraform"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_average", "tags.3241999189", "source:circonus"),
-
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_average", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_average", "unit", "seconds"),
 
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_maximum", "name", metricMaxName),
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_maximum", "active", "true"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_maximum", "tags.#", "2"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_maximum", "tags.2087084518", "author:terraform"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_maximum", "tags.3241999189", "source:circonus"),
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_maximum", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_maximum", "unit", "seconds"),
 
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_minimum", "name", metricMinName),
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_minimum", "active", "true"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_minimum", "tags.#", "0"),
 					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_minimum", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.icmp_ping_minimum", "unit", ""),
 				),
 			},
 		},
@@ -63,7 +52,6 @@ func TestAccCirconusMetric_tagsets(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("circonus_metric.t", "name", metricName),
 					resource.TestCheckResourceAttr("circonus_metric.t", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.#", "0"),
 				),
 			},
 			{
@@ -71,8 +59,6 @@ func TestAccCirconusMetric_tagsets(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("circonus_metric.t", "name", metricName),
 					resource.TestCheckResourceAttr("circonus_metric.t", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.#", "1"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.1750285118", "foo:bar"),
 				),
 			},
 			{
@@ -80,9 +66,6 @@ func TestAccCirconusMetric_tagsets(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("circonus_metric.t", "name", metricName),
 					resource.TestCheckResourceAttr("circonus_metric.t", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.#", "2"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.1750285118", "foo:bar"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.2693443894", "foo:baz"),
 				),
 			},
 			{
@@ -90,10 +73,6 @@ func TestAccCirconusMetric_tagsets(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("circonus_metric.t", "name", metricName),
 					resource.TestCheckResourceAttr("circonus_metric.t", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.#", "3"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.1750285118", "foo:bar"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.2693443894", "foo:baz"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.1937518738", "foo:bur"),
 				),
 			},
 			{
@@ -101,11 +80,6 @@ func TestAccCirconusMetric_tagsets(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("circonus_metric.t", "name", metricName),
 					resource.TestCheckResourceAttr("circonus_metric.t", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.#", "4"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.1750285118", "foo:bar"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.2693443894", "foo:baz"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.1937518738", "foo:bur"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.2110890696", "foo:baz2"),
 				),
 			},
 			{
@@ -113,10 +87,6 @@ func TestAccCirconusMetric_tagsets(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("circonus_metric.t", "name", metricName),
 					resource.TestCheckResourceAttr("circonus_metric.t", "type", "numeric"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.#", "3"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.1750285118", "foo:bar"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.1937518738", "foo:bur"),
-					resource.TestCheckResourceAttr("circonus_metric.t", "tags.2110890696", "foo:baz2"),
 				),
 			},
 		},
@@ -147,24 +117,12 @@ resource "circonus_metric" "icmp_ping_average" {
   name = "%s"
   active = false
   type = "numeric"
-  unit = "seconds"
-
-  tags = [
-    "author:terraform",
-    "source:circonus",
-  ]
 }
 
 resource "circonus_metric" "icmp_ping_maximum" {
   name = "%s"
   active = true
   type = "numeric"
-  unit = "seconds"
-
-  tags = [
-    "source:circonus",
-    "author:terraform",
-  ]
 }
 
 resource "circonus_metric" "icmp_ping_minimum" {
@@ -176,12 +134,6 @@ resource "circonus_metric" "icmp_ping_minimum" {
 const testAccCirconusMetricTagsFmt0 = `
 resource "circonus_metric" "t" {
   name = "%s"
-# tags = [
-#    "foo:bar",
-#    "foo:baz",
-#    "foo:bur",
-#    "foo:baz2"
-# ]
   type = "numeric"
 }
 `
@@ -189,12 +141,6 @@ resource "circonus_metric" "t" {
 const testAccCirconusMetricTagsFmt1 = `
 resource "circonus_metric" "t" {
   name = "%s"
-  tags = [
-    "foo:bar",
-#    "foo:baz",
-#    "foo:bur",
-#    "foo:baz2"
-  ]
   type = "numeric"
 }
 `
@@ -202,12 +148,6 @@ resource "circonus_metric" "t" {
 const testAccCirconusMetricTagsFmt2 = `
 resource "circonus_metric" "t" {
   name = "%s"
-  tags = [
-    "foo:bar",
-    "foo:baz",
-#    "foo:bur",
-#    "foo:baz2"
-  ]
   type = "numeric"
 }
 `
@@ -215,12 +155,6 @@ resource "circonus_metric" "t" {
 const testAccCirconusMetricTagsFmt3 = `
 resource "circonus_metric" "t" {
   name = "%s"
-  tags = [
-    "foo:bar",
-    "foo:baz",
-    "foo:bur",
-#    "foo:baz2"
-  ]
   type = "numeric"
 }
 `
@@ -228,12 +162,6 @@ resource "circonus_metric" "t" {
 const testAccCirconusMetricTagsFmt4 = `
 resource "circonus_metric" "t" {
   name = "%s"
-  tags = [
-    "foo:bar",
-    "foo:baz",
-    "foo:bur",
-    "foo:baz2"
-  ]
   type = "numeric"
 }
 `
@@ -241,12 +169,6 @@ resource "circonus_metric" "t" {
 const testAccCirconusMetricTagsFmt5 = `
 resource "circonus_metric" "t" {
   name = "%s"
-  tags = [
-    "foo:bar",
-#    "foo:baz",
-    "foo:bur",
-    "foo:baz2"
-  ]
   type = "numeric"
 }
 `