+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 0.10.0 (Unreleased)
+
+FIXES:
+
+* remove support for deprecated metric.tags and metric.units (use streamtags)
+* remove support for deprecated metric_cluster
+
 ## 0.9.0 (April 03, 2020)
 
 IMPROVEMENTS: